feat: create endpoints to store subscription and plans

+++ models/subscription.go
@@ -11,7 +11,7 @@ type Subscription struct {
 	Name        string         `json:"name"         gorm:"type:varchar(100);not null"`
 	Price       float32        `json:"price"        gorm:"type:decimal(10,2);not null"`
 	Description string         `json:"description"  gorm:"type:text"`
-	Duration    int8           `json:"duration"     gorm:"type:int;not null"`
+	Duration    int16          `json:"duration"     gorm:"type:int;not null"`
 	CreatedAt   time.Time      `json:"created_at"   gorm:"type:timestamp;default:CURRENT_TIMESTAMP"`
 	UpdatedAt   time.Time      `json:"updated_at"   gorm:"type:timestamp;default:CURRENT_TIMESTAMP"`
 	DeletedAt   gorm.DeletedAt `json:"-"            gorm:"type:timestamp;index"`
